refactor: rename discussion_handler to discussion_orchestrator (#321)
Consistent with coding_orchestrator and orchestrator naming convention.
