+++ docs/adr/0007-yaml-mapping-credential-registry.md
@@ -43,4 +43,4 @@ Option B. The YAML → rendered JSON pipeline gives us human-friendly authoring 
 - **Risk**: New env vars (`GITLAB_TOKEN__*`) require container restart to take effect — only mapping changes are hot-reloadable.
 - **Follow-up needed**: Secondary code paths (`gitlab_poller.py`) still use the global `GITLAB_TOKEN`. Per-project tokens only flow through the Jira polling → coding pipeline.
 
-> **Note**: `mr_comment_handler.py` referenced in the original ADR was replaced by `discussion_handler.py`.
+> **Note**: `mr_comment_handler.py` referenced in the original ADR was replaced by `discussion_orchestrator.py`.
